+++ astrocook/gui_dialog.py
@@ -198,7 +198,7 @@ class GUIDialogMethod(GUIDialog):
         self._box_buttons(self._cancel_run)
         self.SetSizer(self._bottom)
         self.Centre()
-        self.SetPosition((self.GetPosition()[0], wx.DisplaySize()[1]*0.25))
+        self.SetPosition((self.GetPosition()[0], int(wx.DisplaySize()[1]*0.25)))
         self.Show()
 
 
@@ -257,7 +257,7 @@ class GUIDialogMethods(GUIDialog):
         self._box_buttons()
         self.SetSizer(self._bottom)
         self.Centre()
-        self.SetPosition((self.GetPosition()[0], wx.DisplaySize()[1]*0.25))
+        self.SetPosition((self.GetPosition()[0], int(wx.DisplaySize()[1]*0.25)))
         self.Show()
 
     def _box_methods(self):
@@ -300,7 +300,7 @@ class GUIDialogMini(wx.Dialog):
         self._box_buttons()
         self.SetSizer(self._bottom)
         self.Centre()
-        self.SetPosition((self.GetPosition()[0], wx.DisplaySize()[1]*0.25))
+        self.SetPosition((self.GetPosition()[0], int(wx.DisplaySize()[1]*0.25)))
         self.Show()
